refactor: 完善打包
