+++ vite.config.js
@@ -1,7 +1,8 @@
 import { resolve, dirname } from "path";
-import { defineConfig } from "vite";
-
 import { fileURLToPath } from "node:url";
+
+import { defineConfig } from "vite";
+import dts from "vite-plugin-dts";
 
 export default defineConfig({
   build: {
@@ -14,4 +15,5 @@ export default defineConfig({
       output: {},
     },
   },
+  plugins: [dts()],
 });
